+++ com.unity.shadergraph/Editor/Data/Nodes/Utility/CustomFunctionNode.cs
@@ -38,7 +38,7 @@ namespace UnityEditor.ShaderGraph
 
         private static string m_DefaultFunctionName = "Enter function name here...";
 
-        public string functionName 
+        public string functionName
         {
             get => m_FunctionName;
             set => m_FunctionName = value;
@@ -86,7 +86,7 @@ namespace UnityEditor.ShaderGraph
                 }
                 return;
             }
-            
+
             foreach (var argument in slots)
                 visitor.AddShaderChunk(string.Format("{0} _{1}_{2};",
                     NodeUtils.ConvertConcreteSlotValueTypeToString(precision, argument.concreteValueType),
@@ -94,7 +94,7 @@ namespace UnityEditor.ShaderGraph
 
             string call = string.Format("{0}_{1}(", functionName, precision);
             bool first = true;
-            
+
             slots.Clear();
             GetInputSlots<MaterialSlot>(slots);
             foreach (var argument in slots)
@@ -220,12 +220,31 @@ namespace UnityEditor.ShaderGraph
             }
         }
 
+        void ValidateSlotName()
+        {
+            List<MaterialSlot> slots = new List<MaterialSlot>();
+            GetSlots(slots);
+
+            foreach (var slot in slots)
+            {
+                var error = NodeUtils.ValidateSlotName(slot.RawDisplayName(), out string errorMessage);
+                if (error)
+                {
+                    owner.AddValidationError(tempId, errorMessage);
+                    break;
+                }
+            }
+        }
+
         public override void ValidateNode()
         {
             if (!this.GetOutputSlots<MaterialSlot>().Any())
             {
                 owner.AddValidationError(tempId, s_MissingOutputSlot, ShaderCompilerMessageSeverity.Warning);
             }
+            
+            ValidateSlotName();
+            
             if(sourceType == HlslSourceType.File)
             {
                 if(!string.IsNullOrEmpty(functionSource))
@@ -244,7 +263,7 @@ namespace UnityEditor.ShaderGraph
             
             base.ValidateNode();
         }
-        
+
         public override void GetSourceAssetDependencies(List<string> paths)
         {
             base.GetSourceAssetDependencies(paths);
@@ -255,7 +274,7 @@ namespace UnityEditor.ShaderGraph
                     paths.Add(dependencyPath);
             }
         }
-        
+
         public VisualElement CreateSettingsElement()
         {
             PropertySheet ps = new PropertySheet();